# AgriTrace.API/swagger.yaml
openapi: 3.0.4
info:
  title: AgriTrace API — Agricultural Supply Chain Traceability System
  description: "REST API for the Agricultural Supply Chain Traceability System.\n\n**ID Type Note:** organizationId, categoryId, productId, userId are Guid (UUID) in this implementation. The swagger.yaml spec declares them as integer. This is a documented architectural decision to preserve domain integrity."
  contact:
    name: AgriTrace Team
  version: '1.0'
paths:
...
  /api/v1/products:
    get:
      tags:
        - Products
      summary: Danh sách Product
      parameters:
        - name: organizationId
          in: query
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags:
        - Products
      summary: Tạo Product mới
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/v1/products/{productId}':
    get:
      tags:
        - Products
      summary: Chi tiết Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Products
      summary: Cập nhật Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Products
      summary: Xóa Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/v1/products/{productId}/status':
    patch:
      tags:
        - Products
      summary: Thay đổi trạng thái Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductStatusRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ProductStatusRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProductStatusRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/v1/products/{productId}/images':
    get:
      tags:
        - Products
      summary: Danh sách ảnh sản phẩm
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags:
        - Products
      summary: Upload ảnh sản phẩm
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - File
              type: object
              properties:
                File:
                  type: string
                  format: binary
                IsPrimary:
                  type: boolean
            encoding:
              File:
                style: form
              IsPrimary:
                style: form
      responses:
        '501':
          description: Not Implemented
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/v1/products/images/{imageId}':
    delete:
      tags:
        - Products
      summary: Xóa ảnh sản phẩm
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '501':
          description: Not Implemented
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      additionalProperties: false
      description: "Envelope chuẩn cho response thành công của API, khớp schema `ApiResponse` trong swagger.yaml.\r\nHình dạng: { success, data, message, timestamp }."
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
          nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          nullable: true
        timestamp:
          type: string
          format: date-time
      additionalProperties: false
      description: "Envelope chuẩn cho response lỗi của API, khớp schema `ErrorResponse` trong swagger.yaml.\r\nHình dạng: { success, data, message, errors[], timestamp }."
    FieldError:
      type: object
      properties:
        field:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      additionalProperties: false
      description: "Mô tả một lỗi ở cấp field, khớp schema `FieldError` trong swagger.yaml.\r\nHình dạng: { field, code, message }."
...
    ProductRequest:
      required:
        - name
      type: object
      properties:
        name:
          maxLength: 200
          minLength: 0
          type: string
        categoryId:
          type: string
          format: uuid
          nullable: true
        unit:
          maxLength: 50
          minLength: 0
          type: string
          description: Unit code/name (e.g. "kg").
          nullable: true
        unitId:
          type: string
          description: Direct Unit ID (Guid).
          format: uuid
          nullable: true
        organizationId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
      description: "Request body for creating/updating a product.\r\nSupports both string Unit and UUID UnitId for frontend flexibility."
    ProductStatus:
      enum:
        - Created
        - Active
        - Inactive
      type: string
    ProductStatusRequest:
      required:
        - status
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ProductStatus'
      additionalProperties: false
      description: Request body for setting a product's status.
...
  securitySchemes:
    BearerAuth:
      type: http
      description: Enter JWT Bearer token
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: [ ]
...